feat: Create Profile Page
Developed the Profile page featuring a comprehensive introduction to João Pastoriza. The page starts with a HeaderComponent for consistent navigation. It showcases two profile images at the top and bottom, adding a personal touch. The main content is an extensive article detailing João's journey from a young cinema enthusiast to a professional in the field. It highlights his experiences in various aspects of filmmaking and production, his academic journey, and personal interests like football and writing. The page also includes links to download his resume and for contact purposes, facilitating professional connections. The layout is designed for an engaging storytelling experience, with a focus on readability and user engagement.