feat: Add classes for custom image positioning in Headlines
Added `object-top` class to Headline components on the Home page to improve image visualization. This change aims to enhance the aesthetic of the images across different devices, ensuring the most relevant part of the image is always visible.

refactor: Update project IDs in the Home page array

Refactored the array of project IDs on the Home page to reflect recent changes in project identifiers. This ensures that links to individual projects are correct and functional, improving navigation and user experience.